+++ application/Particle/ParticleTest.cpp
@@ -9,7 +9,7 @@ ParticleTest::ParticleTest() :
 	texture = *TextureManager::GetInstance()->GetTexture("uv");
 }
 
-void ParticleTest::Add(size_t addNum, size_t time, Vector3 pos, Vector3 addVec, float scale)
+void ParticleTest::Add(int32_t addNum, int32_t time, Vector3 pos, Vector3 addVec, float scale)
 {
 	for (int i = 0; i < addNum; i++)
 	{
