+++ engine/2d/IPostEffect.h
@@ -35,7 +35,7 @@ private:
 	D3D12_INDEX_BUFFER_VIEW ibView_{};
 	//定数バッファ用データ構造体
 	struct ConstBufferData {
-		size_t timer;
+		int32_t timer;
 	};
 protected:
 	ComPtr<ID3D12Resource> constBuff_ = nullptr;
